+++ core/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for Delphi 13 `noreturn` directive.
 - Added `KeywordKind::is_generic_constraint`.
+- Added support for Delphi 13 `unmanaged` generic constraint.
 
 ## 0.7.0 - 2025-11-11